+++ NEWS.md
@@ -1,5 +1,8 @@
 # odbc (development version)
 
+* Databricks: Improve performance when `useNativeQuery`
+  connection attribute is set to `true` (#998).
+
 * Databricks: Update driver location discovery to handle
   new OEM driver for MacOS (#983).